Removed whitespace from date_string
On Safari we witnessed a bug with the date object being interpreted as invalid. This was because of the whitespace i.e `new Date(date_string + ' ')`. All the beads would default to `date = 0`. Removing this whitespace fixed the bug. Must check whether removing this whitespace raises an issue in Chrome.